% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_plans.R
\name{smGenerateModellingPlan}
\alias{smGenerateModellingPlan}
\title{Code wrapper to generate the modelling plan}
\usage{
smGenerateModellingPlan()
}
\description{
Code wrapper to generate the modelling plan
}
